its a done deal
# Conflicts:
#	src/main/kotlin/link/portalbox/pluginportal/command/sub/DeleteSubCommand.kt
