+++ src/main/kotlin/link/portalbox/pluginportal/command/sub/ListSubCommand.kt
@@ -3,6 +3,7 @@ package link.portalbox.pluginportal.command.sub
 import link.portalbox.pluginportal.command.SubCommand
 import link.portalbox.pluginportal.file.Data
 import link.portalbox.pluginportal.util.ChatColor.color
+import link.portalbox.pluginportal.util.ChatColor.colorOutput
 import link.portalbox.pplib.manager.MarketplaceManager
 import org.bukkit.command.CommandSender
 
@@ -11,11 +12,11 @@ class ListSubCommand : SubCommand() {
     override fun execute(sender: CommandSender, args: Array<out String>) {
         val installedPlugins = Data.installedPlugins
         if (installedPlugins.isEmpty()) {
-            sender.sendMessage("&7&l[&b&lPP&7&l] &8&l> &7You have no plugins installed or recognised by Plugin Portal.".color())
+            sender.sendMessage("&7You have no plugins installed or recognised by Plugin Portal.".colorOutput())
             return
         }
 
-        sender.sendMessage("&7&l[&b&lPP&7&l] &8&l> &7Listing all installed plugins...".color())
+        sender.sendMessage("&7Listing all installed plugins...".colorOutput())
         for (plugin in installedPlugins) {
             sender.sendMessage("&a+ &b${MarketplaceManager.getName(plugin.id)}".color())
         }
